Actualizar index.js

+++ index.js
@@ -1,13 +1,13 @@
 import { join, dirname } from 'path'
 import { createRequire } from 'module'
 import { fileURLToPath } from 'url'
-import cluster from 'cluster'
 import { watchFile, unwatchFile } from 'fs'
 import cfonts from 'cfonts'
 import { createInterface } from 'readline'
 import yargs from 'yargs'
 import chalk from 'chalk'
 import os from 'os'
+import cluster from 'cluster'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 const require = createRequire(__dirname)
@@ -59,44 +59,47 @@ async function start(files) {
   for (const file of files) {
     let args = [join(__dirname, file), ...process.argv.slice(2)]
 
-    cluster.setupMaster({
-      exec: args[0],
-      args: args.slice(1)
-    })
-
-    let p = cluster.fork()
-
-    p.on('message', data => {
-      switch (data) {
-        case 'reset':
-          p.process.kill()
-          isRunning = false
-          start(files)
-          break
-        case 'uptime':
-          p.send(process.uptime())
-          break
-      }
-    })
-
-    p.on('exit', (_, code) => {
-      isRunning = false
-      console.error(chalk.red('❌ Error inesperado:'), code)
-      start(files)
-
-      if (code === 0) return
-      watchFile(args[0], () => {
-        unwatchFile(args[0])
-        start(files)
+    // Nueva API de cluster para Node.js 22+
+    if (cluster.isPrimary || cluster.isMaster) {
+      cluster.setupPrimary({
+        exec: args[0],
+        args: args.slice(1)
       })
-    })
 
-    let opts = new Object(yargs(process.argv.slice(2)).exitProcess(false).parse())
-    if (!opts['test'])
-      if (!rl.listenerCount())
-        rl.on('line', line => {
-          p.emit('message', line.trim())
+      let p = cluster.fork()
+
+      p.on('message', data => {
+        switch (data) {
+          case 'reset':
+            p.process.kill()
+            isRunning = false
+            start(files)
+            break
+          case 'uptime':
+            p.send(process.uptime())
+            break
+        }
+      })
+
+      p.on('exit', (_, code) => {
+        isRunning = false
+        console.error(chalk.red('❌ Error inesperado:'), code)
+        start(files)
+
+        if (code === 0) return
+        watchFile(args[0], () => {
+          unwatchFile(args[0])
+          start(files)
         })
+      })
+
+      let opts = new Object(yargs(process.argv.slice(2)).exitProcess(false).parse())
+      if (!opts['test'])
+        if (!rl.listenerCount())
+          rl.on('line', line => {
+            p.emit('message', line.trim())
+          })
+    }
   }
 }
